Change element_t to uint8_t
Using uint16_t was a silly hack for dealing with PICs, but it never really worked and was far more trouble than it was ever worth.

+++ include/intelhex.h
@@ -24,7 +24,7 @@ namespace intelhex
     {
 	//Each line of the hex file generates a block of memory at a particular address
 	// pair<>.first is the address, pair<>.second is the data
-	typedef	uint16_t	element_t;								//Data element
+	typedef	uint8_t	    element_t;								//Data element
 	typedef	uint32_t	address_t;
 	typedef	std::vector<element_t>	data_container;		//Element container
 	typedef	std::pair<address_t, data_container>	dblock;	//Data block container
